refector: created folder structure
