+++ src/core/decoration.ts
