+++ src/core/enums.ts
